setup snapshots screen
